Fix a spelling mistake and merge duplicate test files
